+++ min_version_test/lib/main.dart
@@ -19,7 +19,7 @@ Future<void> main() async {
         SentryScreenshotWidget(
           child: SentryUserInteractionWidget(
             child: DefaultAssetBundle(
-              bundle: SentryAssetBundle(enableStructuredDataTracing: true),
+              bundle: SentryAssetBundle(),
               child: const MyApp(),
             ),
           ),
@@ -37,14 +37,12 @@ Future<void> setupSentry(AppRunner appRunner) async {
     options.sendDefaultPii = true;
     options.reportSilentFlutterErrors = true;
     options.enableNdkScopeSync = true;
-    options.enableUserInteractionTracing = true;
     options.attachScreenshot = true;
     options.attachViewHierarchy = true;
     // We can enable Sentry debug logging during development. This is likely
     // going to log too much for your app, but can be useful when figuring out
     // configuration issues, e.g. finding out why your events are not uploaded.
     options.debug = true;
-    options.captureFailedHttpRequests = true;
   },
       // Init your App.
       appRunner: appRunner);